Add the go code.
This commit contains our first route. Which is teacher/auth.

Now to the interesting part, we will use negroni as our framework
and not martini. After searching and testing and reading martini isn't
simple enough for go developers. Go in supposed to be KISS and martini isn't.
Negroni from the other hand is KISS and to be honest I liked it more.